shlex: break out the argv parser
